fix(lens): split vazhi's pure frame out of its stateful shell
Move Screen, Row and the Mode/Status types into packages/vazhi/src/Screen.tsx so a rendering test no longer pulls the forward map, ssh and node:net into its import graph; App.tsx keeps only the container.

Agreed by the lowy ⇄ hickey lens debate (finding lowy-6, raised by lowy). Not pushed or merged.
