+++ packages/vazhi/src/App.tsx
@@ -1,19 +1,15 @@
 /**
- * vazhi's screen — one Ink component tree, filling the terminal.
+ * vazhi's stateful shell — the container that owns the forward map, the keys
+ * and the teardown.
  *
- * Ink (React) rather than OpenTUI/SolidJS, which would have matched this repo's
- * usual grain: OpenTUI's renderer is a native Zig core reached over FFI, and it
- * refuses to start on anything below Node 26.4 with `--experimental-ffi`
- * ("OpenTUI native FFI is not available for this runtime yet" — measured on the
- * pinned toolchain's Node 24). The repo's nixpkgs pin tops out at Node 25, so
- * OpenTUI could not run here at all. Ink is node-native, needs no FFI, and owns
- * the two things this screen actually wants from a framework: flexbox layout
- * and redraw-on-resize.
+ * The map's every visible change — a forward opened, cancelled, or LOST on its
+ * own — is a state change of this component, so the map lives here. Quitting is
+ * here too: `q` tears the forwards down and only then lets the process end.
  *
- * The component owns the forward map, because the map's every visible change —
- * a forward opened, cancelled, or LOST on its own — is a state change of this
- * screen. Quitting is here too: `q` tears the forwards down and only then lets
- * the process end.
+ * The frame it renders is `Screen.tsx`, which holds no state at all. The split
+ * is the volatility line: this is the most volatile code in the package (async
+ * opens, process signals, ink's exit), and the frame below it is the most
+ * stable.
  */
 
 import {
@@ -22,28 +18,13 @@ import {
   formatTarget,
   parseTarget,
 } from "@kolu/port-forward";
-import { Box, Text, useApp, useInput, useWindowSize } from "ink";
-import TextInput from "ink-text-input";
+import { Text, useApp, useInput, useWindowSize } from "ink";
 import { useEffect, useMemo, useState } from "react";
-import {
-  clampSelection,
-  formatUptime,
-  forwardUrl,
-  hyperlink,
-  readPromptInput,
-} from "./format.ts";
+import { clampSelection, forwardUrl, readPromptInput } from "./format.ts";
+import { type Mode, Screen, type Status } from "./Screen.tsx";
 
 /** How often the uptime column moves and the list is re-read. */
 const TICK_MS = 1000;
-
-/** The bottom line: the keybind legend, or the `host:port` prompt. */
-type Mode = { kind: "table" } | { kind: "add"; input: string };
-
-/** The one transient message line — what just happened, or what just failed. */
-interface Status {
-  kind: "info" | "error";
-  text: string;
-}
 
 function messageOf(err: unknown): string {
   return err instanceof Error ? err.message : String(err);
@@ -214,128 +195,5 @@ export function App({ hostname }: { hostname: string }) {
         void add(input);
       }}
     />
-  );
-}
-
-/** The frame, as a function of what there is to show — no state, no effects, no
- *  forward map. Every layout claim about vazhi is a claim about this. */
-export function Screen({
-  forwards,
-  hostname,
-  mode,
-  now,
-  onInputChange,
-  onInputSubmit,
-  selected,
-  size,
-  status,
-}: {
-  forwards: readonly Forward[];
-  hostname: string;
-  mode: Mode;
-  now: number;
-  onInputChange: (input: string) => void;
-  onInputSubmit: (input: string) => void;
-  selected: number;
-  size: { columns: number; rows: number };
-  status: Status | undefined;
-}) {
-  const highlighted = clampSelection(selected, forwards.length);
-  // One column width for every target, so the arrows line up whatever the host
-  // names are.
-  const targetWidth = forwards.reduce(
-    (widest, row) => Math.max(widest, formatTarget(row.target).length),
-    0,
-  );
-
-  return (
-    <Box flexDirection="column" width={size.columns} height={size.rows}>
-      <Box>
-        <Text bold>vazhi</Text>
-        <Text dimColor>
-          {` · ${forwards.length === 0 ? "no" : forwards.length} forward${forwards.length === 1 ? "" : "s"} · answering on ${hostname}`}
-        </Text>
-      </Box>
-
-      <Box flexDirection="column" flexGrow={1} marginTop={1}>
-        {forwards.length === 0 ? (
-          <Text dimColor>
-            nothing forwarded yet — press a and type host:port (e.g.
-            pu-dev:5173)
-          </Text>
-        ) : (
-          forwards.map((row, index) => (
-            <Row
-              key={row.key}
-              forward={row}
-              hostname={hostname}
-              now={now}
-              selected={index === highlighted}
-              targetWidth={targetWidth}
-            />
-          ))
-        )}
-      </Box>
-
-      <Box>
-        {status === undefined ? (
-          <Text> </Text>
-        ) : (
-          <Text
-            color={status.kind === "error" ? "red" : undefined}
-            dimColor={status.kind !== "error"}
-          >
-            {status.text}
-          </Text>
-        )}
-      </Box>
-
-      <Box marginTop={1}>
-        {mode.kind === "add" ? (
-          <>
-            <Text>host:port ▸ </Text>
-            <TextInput
-              value={mode.input}
-              placeholder="pu-dev:5173"
-              onChange={onInputChange}
-              onSubmit={onInputSubmit}
-            />
-            <Text dimColor>{"  (enter to add · esc to cancel)"}</Text>
-          </>
-        ) : (
-          <Text dimColor>a add · x cancel · ↑/↓ move · q quit</Text>
-        )}
-      </Box>
-    </Box>
-  );
-}
-
-/** One forward: what it points at, the URL it answers on (a real terminal
- *  hyperlink), and how long it has been up. */
-function Row({
-  forward,
-  hostname,
-  now,
-  selected,
-  targetWidth,
-}: {
-  forward: Forward;
-  hostname: string;
-  now: number;
-  selected: boolean;
-  targetWidth: number;
-}) {
-  const url = forwardUrl(hostname, forward.localPort);
-  const colour = selected ? "cyan" : undefined;
-  return (
-    <Box>
-      <Text color={colour}>{selected ? "› " : "  "}</Text>
-      <Box width={targetWidth} flexShrink={0}>
-        <Text color={colour}>{formatTarget(forward.target)}</Text>
-      </Box>
-      <Text dimColor>{"  →  "}</Text>
-      <Text color={colour}>{hyperlink(url)}</Text>
-      <Text dimColor>{`  up ${formatUptime(now - forward.createdAt)}`}</Text>
-    </Box>
   );
 }
